requirements.txt
streamlit>=1.32.0
crewai[tools]>=0.28.8
litellm>=1.34.0
python-dotenv>=1.0.1
requests>=2.31.0
jinja2>=3.1.3
matplotlib>=3.8.0
networkx>=3.2.1